pollsapi/requirements.txt
Django>=2.0.3
Djangorestframework>=3.7.7
django-rest-swagger>=2.2.0
